+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Unreleased]
 
+### Fixed
+- Fixed syntax error in type hints for map with generic value type.
+
 ## [v1.0.0-beta.5] - 2022-06-07
 
 ### Changed